==> 03 - Scripts/08 Creating function for econometric exercise.Rmd <==
---
title: "Creating functions for econometric exercise"
author: "Daniel Grimaldi / Schar School of Public Policy"
date: "`r Sys.time()`"
output:
  html_document:
    theme: paper
    number_sections: false
bibliography: references.bib
---

```{r setup, echo=TRUE, warning=FALSE, message=FALSE}
if(!require(tidyverse)){install.packages("tidyverse")}
if(!require(magrittr)){install.packages("magrittr")}
if(!require(stargazer)){install.packages("stargazer")}
if(!require(AER)){install.packages("AER")}
if(!require(sandwich)){install.packages("sandwich")}
```

# Creating functions

The first function created in this file will be responsible for calculating the main equation of @grimaldi2021 for different outcome variables, and considering one specific set up for reference years. The main arguments are:

i. wd: a data set as generated by the previous script in this project;
ii. tf: the final year of analysis (or when we intend to measure the effects);
iii. t0: the base year (or when we will set our year of comparison);
iv. tf.lag: the final year for measuring the past trajectory;
v. t0.lag: the start year for measuring the past trajectory

```{r echo=TRUE, warning=FALSE, message=FALSE}
reg_year <- function(wd, tf, t0, tf.lag, t0.lag, cluster){
  
  # munging data set for regressions
  # ihh_merge measures the expected ihh due to the merge
  # ihh and ihh_merge are NA if and only if there is no banking activity at all in the region in 2008.
  # this can be checked by filtering working data to 2008, and running the following command lines:
  # table(is.na(wokring.data$ihh_loans), working.data$n_banks)
  # table(is.na(wokring.data$ihh_merge_loans), working.data$n_banks)
  # For those cases, we will assume LIM to be zero 
  # Following our identifycation strategy, those places would suffer no (direct) impact from the merge
  names(wd) <- gsub(cluster, "cluster", names(wd))
  
  wd %<>% 
    group_by(sector_id) %>% 
    mutate(delta_s_import_br = (first(imports_br_w2[year==tf], default=0) - first(imports_br_w2[year==t0], default=0))/1000,
           delta_s_import_mercosur = (first(imports_mercosur_w2[year==tf], default=0) - first(imports_mercosur_w2[year==t0], default=0))/1000) %>%
    ungroup() %>%
    group_by(sector_id, year) %>%
    mutate(s_emp = sum(s_m_emp),
           s_companies = sum(s_m_companies),
           s_yearly_wage = sum(s_m_yearly_wage),
           s_dec_wage = sum(s_m_dec_wage)) %>% 
    ungroup() %>% 
    mutate(share_s_m_emp = s_m_emp/s_emp,
           share_s_m_companies = s_m_companies/s_companies,
           share_s_m_yearly_wage = s_m_yearly_wage/s_yearly_wage,
           share_s_m_dec_wage = s_m_dec_wage/s_dec_wage) %>% 
    group_by(cluster,.drop = TRUE) %>%
    summarise(delta_emp=sum(s_m_emp[year==tf])-sum(s_m_emp[year==t0]),
              delta_emp_lag=sum(s_m_emp[year==tf.lag])-sum(s_m_emp[year==t0.lag]),
              delta_companies=sum(s_m_companies[year==tf])-sum(s_m_companies[year==t0]),
              delta_companies_lag=sum(s_m_companies[year==tf.lag])-sum(s_m_companies[year==t0.lag]),
              delta_yearly_wage=sum(s_m_yearly_wage[year==tf])-sum(s_m_yearly_wage[year==t0]),
              delta_yearly_wage_lag=sum(s_m_yearly_wage[year==tf.lag])-sum(s_m_yearly_wage[year==t0.lag]),
              delta_dec_wage=sum(s_m_dec_wage[year==tf])-sum(s_m_dec_wage[year==t0]),
              delta_dec_wage_lag=sum(s_m_dec_wage[year==tf.lag])-sum(s_m_dec_wage[year==t0.lag]),
              LIP_br_emp = sum(share_s_m_emp[year==t0] * first(delta_s_import_br), na.rm=TRUE),
              LIP_mercosur_emp = sum(share_s_m_emp[year==t0.lag] * first(delta_s_import_mercosur), na.rm=TRUE),
              LIP_br_companies = sum(share_s_m_companies[year==t0] * first(delta_s_import_br), na.rm=TRUE),
              LIP_mercosur_companies = sum(share_s_m_companies[year==t0.lag] * first(delta_s_import_mercosur), na.rm=TRUE),
              LIP_br_yearly_wage = sum(share_s_m_yearly_wage[year==t0] * first(delta_s_import_br), na.rm=TRUE),
              LIP_mercosur_yearly_wage = sum(share_s_m_yearly_wage[year==t0.lag] * first(delta_s_import_mercosur), na.rm=TRUE),
              LIP_br_dec_wage = sum(share_s_m_dec_wage[year==t0] * first(delta_s_import_br), na.rm=TRUE),
              LIP_mercosur_dec_wage = sum(share_s_m_dec_wage[year==t0.lag] * first(delta_s_import_mercosur), na.rm=TRUE),
              LIM_deposits = first(ihh_merge_deposits[year==2008]) - first(ihh_deposits[year==2008]),
              LIM_loans = first(ihh_merge_loans[year==2008]) - first(ihh_loans[year==2008]),
              LIM_other_loans = first(ihh_merge_other_loans[year==2008]) - first(ihh_loans[year==2008]),
              LIMn_deposits = first(ihhN_merge_deposits[year==2008]) - first(ihhN_deposits[year==2008]),
              LIMn_loans = first(ihhN_merge_loans[year==2008]) - first(ihhN_loans[year==2008]),
              LIM_other_loans = first(ihhN_merge_other_loans[year==2008]) - first(ihhN_other_loans[year==2008]),
              emp0 = sum(s_m_emp[year==t0]),
              companies0 = sum(s_m_companies[year==t0]),
              yearly_wage0 = sum(s_m_yearly_wage[year==t0]),
              dec_wage0 = sum(s_m_dec_wage[year==t0]),
              uf = first(uf)) %>% 
    mutate(across(starts_with("LIM"), ~replace_na(., 0)),
           across(starts_with("delta"), ~./1000)) %>% 
    arrange(cluster)
  
  # defining main regressors
  y.names <- c("emp", "companies", "dec_wage")
  
  # creating regression functions
  ## OLS
  apply_ols <- function(var, data=wd){
    formula.ols <- paste0("delta_", var, " ~ 0 + LIP_br_", var, "*LIM_loans + delta_", var, "_lag + as.factor(uf)")
    reg.ols <- lm(formula.ols,  data=data)
    reg.ols}
  ## IV
  apply_iv <- function(var, data=wd){
    formula.iv <- paste0("delta_", var, " ~ 0 + LIP_br_", var, "*LIM_loans + delta_", var, "_lag + as.factor(uf) | LIP_mercosur_", var, "*LIM_loans + delta_", var, "_lag + as.factor(uf)")
    reg.iv <- ivreg(formula.iv, data=data)
    reg.iv}
  ## IV 1st stage
  apply_1st <- function(var, data=wd){
    formula.1st <- paste0("LIP_br_", var, " ~ LIP_mercosur_", var)
    reg.1st <- lm(formula.1st, data=data)
    reg.1st}
  ## IV reduced form
  apply_rform <- function(var, data=wd){
    formula.rform <- paste0("delta_", var, " ~ 0 + LIP_mercosur_", var, "*LIM_loans + delta_", var, "_lag + as.factor(uf)")
    reg.rform <- lm(formula.rform, data=data)
    reg.rform}
  
  # creating function for adjusting standard errors
  adj.sd <- function(model.list){
    cov.rob.list <- lapply(ols, vcovHC, type="HC1")
    diag.se.list <- lapply(cov.rob.list, diag)  
    robust.se.list <- lapply(diag.se.list, sqrt)
    robust.se.list
  }
  
  # implementing regressions
  ols <- lapply(y.names, apply_ols)
  names(ols) <- y.names
  
  iv <- lapply(y.names, apply_iv)
  names(iv) <- y.names
  
  iv.1st <- lapply(y.names, apply_1st)
  names(iv.1st) <- y.names
  
  iv.reduced <- lapply(y.names, apply_rform)
  names(iv.reduced) <- y.names
  
  # calculating robust standard errors
  ## OLS
  ols.sd <- adj.sd(ols)
  ## IV
  iv.sd <- adj.sd(iv)
  ## IV 1st stage
  iv.1st.sd <- adj.sd(iv.1st)
  ## IV Reduced form
  iv.reduced.sd <- adj.sd(iv.reduced)
  
  output <- list(data=wd, reg=list(ols=ols, iv=iv, iv.1st=iv.1st, iv.reduced=iv.reduced), sd=list(ols=ols.sd, iv=iv.sd, iv.1st=iv.1st.sd, iv.reduced=iv.reduced.sd))
  output
}
```

The second function will use the previous one to generate a series of estimative, considering a range for tf. The results (beta and sd) are consolidated into a long shape table.

```{r echo=TRUE, warning=FALSE, message=FALSE}
reg_range <- function(tf.range, t0, tf.lag, t0.lag, wd, cluster){
  
  reg.range <- lapply(tf.range, reg_year, t0=t0, tf.lag=tf.lag, t0.lag=t0.lag, wd=wd, cluster=cluster)
  names(reg.range) <- tf.range

  table <- expand.grid(tf.range, c("ols", "iv"), c("emp", "companies", "dec_wage"), c("LIP", "LIM", "LIP:LIM"))
  names(table) <- c("year", "model", "outcome.var", "coef")
  table %<>%
    mutate(across(c("year", "model", "outcome.var", "coef"), ~as.character(.)),
           center = NA,
           sd = NA,
           lower = NA,
           upper = NA)
  
  for (i in 1:nrow(table)){
    model.i <- summary(reg.range[[table$year[i]]][["reg"]][[table$model[i]]][[table$outcome.var[i]]])$coefficients
    coef <- case_when(table$coef[i]=="LIP" ~ row.names(model.i)[1],
                      table$coef[i]=="LIM" ~ row.names(model.i)[2],
                      table$coef[i]=="LIP:LIM" ~ row.names(model.i)[4]) 
    table$center[i] <- model.i[coef, "Estimate"]
    table$sd[i] <- model.i[coef,"Std. Error"]
    table$lower[i] <- table$center[i] - 1.96*table$sd[i]
    table$upper[i] <- table$center[i] + 1.96*table$sd[i]
  }
  output <- list(regs=reg.range, summary=table)
  output
}
```

# Using functions for analysis

## Case 1: microregions considering only manufacturing jobs

```{r echo=TRUE, warning=FALSE, message=FALSE}
load("../02 - Data sets/working_data_microregion_manufacturing.Rdata")
working.data %<>%
  mutate(uf=substr(microregion_id, 1, 2))

analysis_micro <- reg_range(tf.range=2001:2018, t0=2000, tf.lag=1999, t0.lag=1995, wd=working.data, cluster="microregion_id")
filter(analysis_micro$summary, model=="iv") %>% 
  select(-sd)
```
